Moving `get_ran_qasm` test into the `test_wrapper.py` test suite.
