--- test/python/test_wrapper_helpers.py
@@ -1,54 +1,0 @@
-# -*- coding: utf-8 -*-
-
-# Copyright 2018, IBM.
-#
-# This source code is licensed under the Apache License, Version 2.0 found in
-# the LICENSE.txt file in the root directory of this source tree.
-
-# pylint: disable=invalid-name
-
-"""Tests for the wrapper functionality."""
-
-import unittest
-
-from qiskit import QuantumRegister, ClassicalRegister, QuantumCircuit
-from qiskit.wrapper import execute
-from .common import QiskitTestCase, is_cpp_simulator_available
-
-
-class TestWrapperHelpers(QiskitTestCase):
-    """Suite for testing `compile`, `execute` and other execution helpers."""
-
-    def setUp(self):
-        q = QuantumRegister(3)
-        c = ClassicalRegister(3)
-        self.circuit = QuantumCircuit(q, c)
-        self.circuit.ccx(q[0], q[1], q[2])
-        self.circuit.measure(q, c)
-
-    def test_local_execute_and_get_ran_qasm(self):
-        """Check if the local backend return the ran qasm."""
-
-        cpp_simulators = [
-            'local_qasm_simulator_cpp',
-            'local_statevector_simulator_cpp'
-        ]
-
-        python_simulators = [
-            'local_qasm_simulator_py',
-            'local_statevector_simulator_py',
-            'local_unitary_simulator_py'
-        ]
-
-        local_simulators = python_simulators
-        if is_cpp_simulator_available():
-            local_simulators += cpp_simulators
-
-        for backend_name in local_simulators:
-            with self.subTest(backend_name=backend_name):
-                result = execute(self.circuit, 'local_qasm_simulator').result()
-                self.assertIsNotNone(result.get_ran_qasm(self.circuit.name))
-
-
-if __name__ == '__main__':
-    unittest.main(verbosity=2)
